Moved Animal babies to Ageable entities. Added Baby/Parent implementation.
